@sam-s 2016-01-21
I don't think making this change optional makes much sense.
Fast predict should always be on and if the user wants to preserve the training data, they should do it on their own.