@amueller 2016-01-21
I agree, making this optional is weird. It looks like a great addition.
We could either change ``X_`` directly, which might break people's code (bad) or try a work-around.
We could add ``store_X=None`` and have it be ``True`` by default but deprecate that and announce changing the default value?
But as it only touches the attributes, I think it might be better to add a deprecated property.
Maybe renaming ``X_`` and ``y_`` makes more sense, and having the old ones be deprecated copies?

New names might be ``fit_X_`` and ``isotonic_y_``? I don't know.
We would need to store ``fit_X_`` for two versions for backward compatibility, but everyone would have fast predictions.