@jarfa 2016-01-21
Interesting idea re:fit_X and isotonic_y, but are we concerned at all about storage? The reason I made this change in the first place is that we were trying to calibrate with a large training set.