@sam-s 2016-01-21
To be quite frank, I can live with breaking user code which relies on internals.