@TomDLT 2016-01-22
You will have to add a test to `sklearn.tests.test_isotonic.py` to be sure it does not change the predictions.