@amueller 2016-01-22
@sam-s it's not really internals. It's a documented attribute. Therefore it's part of the API. If it was a private attribute (starting with ``_``) that would be different.
We could also just make the new ones private ``_X`` and ``_y``.
I agree about storage, but I think we have to store it for two versions until we can remove it.
The only way not to store it is to add another parameter, which we have to remove after the storage option got removed. So that means this will create changes four versions from now.