@jarfa 2016-01-25
Made a bunch of changes, hopefully the automated checks won't throw any errors. When I created my test function I noticed that it was failing because ~20% of samples were different between the new and old methods - however, their differences were all on the order of 1e-16. After a bit of investigation I found out that this was because interpolation for prediction wasn't linear - it's a 1st-order spline function.

I changed the interpolation to linear because:  
a) this makes more sense if we're removing the points in the middle of a flat interval   
b) the difference is negligible - again, it's on the order of 1e-16

I left fast_predict as an option for now simply because it's necessary for the test script to prove that this change works. I'm happy to make it the default option, to add the training data as private self._x & self._y... or whatever you all want. Or you all could merge these changes and make the faster prediction fit into your grander scheme of sci-kit learn, I am after all new to contributing to this project.

Thanks!