@jarfa 2016-01-25
Apparently sklearn only changed from linear to spline interpolation this month (https://github.com/scikit-learn/scikit-learn/commit/a9ea55fb404ba3a2411afa7a329f69eda67ca86c). The pull request didn't say much about why the change was made.