@jarfa 2016-01-25
@agramfort : I fixed the white-space issues you pointed out. Your point about making the parameter part of init instead of fit is valid, but from the above conversation with @amueller it sounds like it's not even going to be an option. 

Still waiting for @amueller to opine on how to handle this (should we call the new self.X_ and self.y_  self.fit_X_ and self.isotonic_y_ ?)