@amueller 2016-01-25
I think adding new private variables and deprecated properties for backward compatibility is the way to go.