@jarfa 2016-01-25
Done. I'm calling them self._necessary_X_ and self.necessary_y_.