@agramfort 2016-01-26
you should also raise deprecation warnings when trying to access X_ and y_ attributes. You could do it with properties