@ogrisel 2016-01-28
See:

http://scikit-learn.org/stable/developers/contributing.html#deprecation