@ogrisel 2016-01-28
Also @jarfa if you are proficient with git, please squash the commits of this PR. If you don't know how to do that, don't worry, the person that will merge this PR will do it prior to merging.