@jarfa 2016-01-29
@ogrisel thanks, that's good feedback. I'll make sure the checks pass, then I'll get to squashing some commits.