@jarfa 2016-01-29
@ogrisel Integrated your changes and squashed them all into 1.