@agramfort 2016-01-30
can you update what's new?