@jarfa 2016-01-31
@agramfort, are you just asking for a run-down of the most recent changes?

- replaced the attributes `self.X_` and `self.y_` with properties that throw deprecation warnings ("... is deprecated in version 0.18 and will be removed in version 0.20."). They refer to attributes  `self._X_` and `self._y_`
- The attributes that hold only the points necessary for prediction are called `self._necessary_X_` and `self._necessary_y_`
- testing no longer has `np.random.seed(...)`, I now create a random state object.