@agramfort 2016-01-31
sorry I meant the doc file whats_new.rst that documents the new features, bug fixes and API changes.