@jarfa 2016-01-31
@agramfort 4/4 tests were passing before the latest commit, and as you can see 1/1 have now failed.

I only changed the whats_new.rst file: how exactly did I mess that up?