@agramfort 2016-01-31
your branch cannot be merged with master due to conflict in what's new file. You need to rebase and fix the conflict.