@jarfa 2016-02-03
I took care of the merge problems and the tests all passed. Aside from the ongoing question about spline vs. linear interpolation, what else needs to be done before merging into master?

One issue - I'm leaving Friday (2/05) morning for a 2-week vacation without my laptop. I'll have email access but not be able to work on code - so after tomorrow, any additional changes will either need to be done by somebody else or will have to wait until Feb. 21st.