@ogrisel 2016-02-04
The deprecated attribute `X_` and `y_` should be removed from the list of documented attributes in the docstring and the new attributes should be documented instead.