@ogrisel 2016-02-04
+1 for merge once my last batch of comments has been addressed including https://github.com/scikit-learn/scikit-learn/pull/6206#discussion_r51859607.