@jarfa 2016-02-04
Thanks for your support @ogrisel, I  made your suggested change to `test_isotonic_regression_ties_min()` and put it in a separate commit. 

Again, after tonight (US Eastern time) I'll be too busy hiking around the Colombian coffee country to do more work on this for the next 2 weeks. Therefore if we need any other changes I'd suggest that you or somebody else clone this branch and do whatever else is necessary.

It's been great making this small contribution to the project, I hope I'll be able to do more in the future.