@ogrisel 2016-02-05
Thanks very much for your contribution @jarfa. I will address my last comment (https://github.com/scikit-learn/scikit-learn/pull/6206#discussion_r51992608) myself and as @tiagozortea made additional checks on his own environment and gave his +1 I will merge consecutively.