@ogrisel 2016-02-05
I merged #6286, closing this. Thanks again for the fix @jarfa and for the tests @tiagozortea.