@mjbommar 2016-02-18
Would like to point us back at this thread re: `fit`/`transform` vs. `fit_transform` and why we should keep `X` and `y`: https://github.com/scikit-learn/scikit-learn/pull/4185#issuecomment-72875303

I am not a fan of fully deprecating access to my case A in comments above; it has many applications in the pre-processing and conditioning of data with known monotonic relationships prior, e.g., in credit scoring models.